No more multi-dimensional relation.
The possibility of a multi-dimensional relation was available in the
settings but not used, because it prevented certain relationships from
being perceived. Removing this multi-dimensional option and defining a
relation as necessarily having a single dimension makes the code much
simpler and opens the possibility of a more precise dimension coding.