New dimension coding procedure
The notion of "dimension" referred to a path in the syntactic tree. This
led to ambiguous situations when the varying value was the parameter of
a relation, ex: "group.relations.relation.sequence.commonDiff". If many
sequences can be found in the component descriptions, which one will see
its common difference vary ?

In order to address this issue, the notion of dimension has been
dissociated from that of path and now refer to a list of successive
paths, containing at least one element, and where the last path leads to
a non-relational value (ex : "size").

A dimension is coded according to this pattern : p.a.t.h.1>p.a.t.h.2>...